Image edit and delete validation

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,20 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
+        String error = "Only the owner of the image can edit the image";
         Image image = imageService.getImage(imageId);
-
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
-        return "images/edit";
+
+        User sessionUser = (User) session.getAttribute("loggeduser");
+        if(sessionUser.getId() != image.getUser().getId()) {
+            model.addAttribute("editError", error);
+            model.addAttribute("tags", image.getTags());
+            return "images/image";
+        }
+            return "images/edit";
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -132,7 +139,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/"+ imageId + "/" + updatedImage.getTitle();
     }
 
 
@@ -140,7 +147,19 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        String error = "Only the owner of the image can delete the image";
+        Image image = imageService.getImage(imageId);
+        String tags = convertTagsToString(image.getTags());
+        model.addAttribute("image", image);
+        model.addAttribute("tags", tags);
+
+        User sessionUser = (User) session.getAttribute("loggeduser");
+        if(sessionUser.getId() != image.getUser().getId()) {
+            model.addAttribute("deleteError", error);
+            model.addAttribute("tags", image.getTags());
+            return "images/image";
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
